inventory_manage/requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.3
django-import-export==4.3.3
django-leaflet==0.31.0
djangorestframework==3.15.2
pillow==11.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.2
tablib==3.7.0
